requirements.txt
tensorflow
gensim
nltk
pandas
numpy
ipykernel
kaggle
python-dotenv
scikit-learn
scipy
fastapi
uvicorn
python-multipart
PyPDF2
jinja2
pyyaml

-e .
